# Clan Update <@&1278105457965404180> - Clans are added and tested - Clan bank - Clan Chat - Clan Roles: `Leader,General,Member` - Clan Exp/Level & Quest are shown (Not 100% Added yet) - Clan Info (Displays ALOT of info) # Misc Update - Criminal's Can't resign. - Huge Refactoring - Added 8 files that are needed Refactoring - New Map is Getting Some progress give a Thx to Built Team! - /hello improved - GitHub will be adding Channel and Webhook / Public soon. - Leaderboards are tested formatted and now only Values need formatting before Give ALL players acess - new Bug Finder Pizzamaniac100 ** <#1267218651765932033> are gonna be accepted/denied and implemented in the next few Days/Weeks! may take a break on Weekends **
# OTHER BUGS/STUFF are ADDED THIS IS UPDATING THEM ALL NOW TO GITHUB

+++ src/main/java/kingsbutbad/kingsbutbad/utils/FormatUtils.java
@@ -15,7 +15,7 @@ public class FormatUtils {
             } else {
                 String[] suffixes = new String[]{"", "K", "M", "B", "T", "P", "E", "Z", "Y"};
                 int magnitude = Math.min((int)Math.log10(bal) / 3, suffixes.length - 1);
-                double scaledValue = bal / Math.pow(1000.0, (double)magnitude);
+                double scaledValue = bal / Math.pow(1000.0, magnitude);
                 NumberFormat formatter = NumberFormat.getCurrencyInstance(Locale.US);
                 formatter.setMaximumFractionDigits(1);
                 return (isNegative ? "-" : "") + formatter.format(scaledValue) + suffixes[magnitude];
